+++ packages/stream_video/lib/src/models/call_session_data.dart
@@ -5,7 +5,7 @@ typedef Role = String;
 typedef UserSessionId = String;
 
 @immutable
-class CallSessionData with EquatableMixin {
+class CallSessionData extends Equatable {
   const CallSessionData({
     this.id = '',
     this.participants = const {},
@@ -42,7 +42,7 @@ class CallSessionData with EquatableMixin {
 }
 
 @immutable
-class CallParticipant with EquatableMixin {
+class CallParticipant extends Equatable {
   const CallParticipant({
     required this.userSessionId,
     required this.userId,